t/data/parsing/mml-1.dat

#data
aa<math>bb<foo>baa</foo>bcc</math>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "aa"
|     <math math>
|       "bb"
|       <math foo>
|         "baa"
|       "bcc"
|     "a"

#data
<matH>aa<altglyph/>cc</math>d
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       "aa"
|       <math altglyph>
|       "cc"
|     "d"

#data
<math definitionurl=abc>aa<x definitionurl=xx repeatdur=1></x>aa</math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       definitionURL="abc"
|       "aa"
|       <math x>
|         definitionURL="xx"
|         repeatdur="1"
|       "aa"

#data
<math xmlns="abc" xlink:tITle="bb" XML:lang="cc">aa</math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       xlink title="bb"
|       xml lang="cc"
|       xmlns xmlns="abc"
|       "aa"

#data
<math><g xmlns="abc" xlink:tITle="bb" XML:lang="cc">aa</g></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math g>
|         xlink title="bb"
|         xml lang="cc"
|         xmlns xmlns="abc"
|         "aa"

#data
<math><p>aaaa</math>
#errors
1;20;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math p>
|         "aaaa"

#data
<math>hoge<br>foo</math>
#errors
1;24;m;not closed;;br
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       "hoge"
|       <math br>
|         "foo"

#data
<math><table><tr>foo
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math table>
|         <math tr>
|           "foo"

#data
<math>foo<p>bar<p>bc</math>
#errors
1;27;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       "foo"
|       <math p>
|         "bar"
|         <math p>
|           "bc"

#data
<math>fo<optgroup>bar<option>aa<optgroup>aa</math>
#errors
1;50;m;not closed;;optgroup
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       "fo"
|       <math optgroup>
|         "bar"
|         <math option>
|           "aa"
|           <math optgroup>
|             "aa"

#data
<math>foo<html>BAR<body>abz</html>aaa</math>
#errors
1;16;m;start tag not allowed;;html
1;25;m;start tag not allowed;;body
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       "fooBARabzaaa"

#data
<math><title>foo<p>bar</p>&amp;</title></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math title>
|         "foo"
|         <math p>
|           "bar"
|         "&"

#data
<math><xmp>foo<p>bar</p>&amp;</xmp></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math xmp>
|         "foo"
|         <math p>
|           "bar"
|         "&"

#data
<math><plaintext>foo<p>bar</p>&amp;</plaintext></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math plaintext>
|         "foo"
|         <math p>
|           "bar"
|         "&"

#data
<math><mi><P>foo</mi></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math mi>
|         <p>
|           "foo"

#data
<math><mo><P>foo</mo></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math mo>
|         <p>
|           "foo"

#data
<math><ms><P>foo</ms></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math ms>
|         <p>
|           "foo"

#data
<math><mn><P>foo</mn></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math mn>
|         <p>
|           "foo"

#data
<math><mtext><P>foo</mText></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math mtext>
|         <p>
|           "foo"

#data
<math><annotation-XML><P>foo</annotation-XML></matH>
#errors
1;46;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         <math p>
|           "foo"

#data
<math><annotation-XML encoding><P>foo</annotation-XML></matH>
#errors
1;55;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding=""
|         <math p>
|           "foo"

#data
<math><annotation-XML encoding=text/xml><P>foo</annotation-XML></matH>
#errors
1;64;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/xml"
|         <math p>
|           "foo"

#data
<math><annotation-XML encoding=text/HTml><P>foo</annotation-XML></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/HTml"
|         <p>
|           "foo"

#data
<math><annotation-XML encoding=
application/xhtml+xml><P>foo</annotation-XML></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="application/xhtml+xml"
|         <p>
|           "foo"

#data
<math><annotation-XML encoding="text/HTml "><P>foo</annotation-XML></matH>
#errors
1;68;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/HTml "
|         <math p>
|           "foo"

#data
<math><annotation-XML encoding=text/HTml><P>foo</annotation-XML></matH>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/HTml"
|         <p>
|           "foo"

#data
<math><annotation-xml encoding="text/html"><svg><g/></svg></annotation-xml></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <svg svg>
|           <svg g>

#data
<math><mi><svg><P>foo</svg></mi></matH>
#errors
1;28;m;not closed;;p
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math mi>
|         <svg svg>
|           <svg p>
|             "foo"

#data
<math><mi><mglyph/>a</mi></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math mi>
|         <math mglyph>
|         "a"

#data
<math><mi><malignmark/>a</mi></math>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math math>
|       <math mi>
|         <math malignmark>
|         "a"

#data
<mglyph><hoge/></mglyph>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math mglyph>
|       <math hoge>

#data
<malignmark><hoge/></malignmark>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <math malignmark>
|       <math hoge>
